services/market-crawler/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
httpx
beautifulsoup4
pydantic
